Clarifying the arguments for addAlias
Our team was slightly confused by the order of the `addAlias` arguments, namely, _which_ alias was then to be used in ability checks. We had interpreted the arguments from left-to-right; for instance: 
I'm defining the string `modify` for use as an alias for the following action rules: `update` and `delete`. When actually, it's the other way around: `update` and `delete` are aliases for the `modify` ability defined in the rules.

I just added a sentence to make it a bit clearer how to use the aliases during checks. :)